add steps for backward compat test

+++ tests/backward_comaptibility_tests.sh
@@ -2,6 +2,7 @@
 set -ev
 
 need_launch=${1:-0}
+start_from=${2:-0}
 
 source ~/.bashrc 
 CLUSTER_NAME="test-back-compat-$USER"
@@ -32,6 +33,7 @@ pip install -e ".[all]"
 
 
 # exec + launch
+if [ "$start_from" -le 1 ]; then
 mamba activate sky-back-compat-master
 rm -r  ~/.sky/wheels || true
 which sky
@@ -48,8 +50,10 @@ echo $s
 # remove color and find the job id
 echo $s | sed -r "s/\x1B\[([0-9]{1,3}(;[0-9]{1,2})?)?[mGK]//g" | grep "Job ID: 3" || exit 1
 sky queue ${CLUSTER_NAME}
+fi
 
 # sky stop + sky start + sky exec
+if [ "$start_from" -le 2 ]; then
 mamba activate sky-back-compat-master
 rm -r  ~/.sky/wheels || true
 sky launch --cloud gcp -y -c ${CLUSTER_NAME}-2 examples/minimal.yaml
@@ -60,19 +64,23 @@ sky start -y ${CLUSTER_NAME}-2
 s=$(sky exec --cloud gcp -d ${CLUSTER_NAME}-2 examples/minimal.yaml)
 echo $s
 echo $s | sed -r "s/\x1B\[([0-9]{1,3}(;[0-9]{1,2})?)?[mGK]//g" | grep "Job ID: 2" || exit 1
+fi
 
 # `sky autostop` + `sky status -r`
+if [ "$start_from" -le 3 ]; then
 mamba activate sky-back-compat-master
 rm -r  ~/.sky/wheels || true
 sky launch --cloud gcp -y -c ${CLUSTER_NAME}-3 examples/minimal.yaml
 mamba activate sky-back-compat-current
 rm -r  ~/.sky/wheels || true
 sky autostop -y -i0 ${CLUSTER_NAME}-3
-sleep 100
+sleep 120
 sky status -r | grep ${CLUSTER_NAME}-3 | grep STOPPED || exit 1
+fi
 
 
 # (1 node) sky launch + sky exec + sky queue + sky logs
+if [ "$start_from" -le 4 ]; then
 mamba activate sky-back-compat-master
 rm -r  ~/.sky/wheels || true
 sky launch --cloud gcp -y -c ${CLUSTER_NAME}-4 examples/minimal.yaml
@@ -85,8 +93,10 @@ sky logs ${CLUSTER_NAME}-4 1 --status
 sky logs ${CLUSTER_NAME}-4 2 --status
 sky logs ${CLUSTER_NAME}-4 1
 sky logs ${CLUSTER_NAME}-4 2
+fi
 
 # (1 node) sky start + sky exec + sky queue + sky logs
+if [ "$start_form" -le 5 ]; then
 mamba activate sky-back-compat-master
 rm -r  ~/.sky/wheels || true
 sky launch --cloud gcp -y -c ${CLUSTER_NAME}-5 examples/minimal.yaml
@@ -101,8 +111,10 @@ sky launch --cloud gcp -y -c ${CLUSTER_NAME}-5 examples/minimal.yaml
 sky queue ${CLUSTER_NAME}-5
 sky logs ${CLUSTER_NAME}-5 2 --status
 sky logs ${CLUSTER_NAME}-5 2
+fi
 
 # (2 nodes) sky launch + sky exec + sky queue + sky logs
+if [ "$start_from" -le 6 ]; then
 mamba activate sky-back-compat-master
 rm -r  ~/.sky/wheels || true
 sky launch --cloud gcp -y -c ${CLUSTER_NAME}-6 examples/multi_hostname.yaml
@@ -117,5 +129,6 @@ sky exec --cloud gcp ${CLUSTER_NAME}-6 examples/multi_hostname.yaml
 sky queue ${CLUSTER_NAME}-6
 sky logs ${CLUSTER_NAME}-6 2 --status
 sky logs ${CLUSTER_NAME}-6 2
+fi
 
 sky down ${CLUSTER_NAME}* -y
